+++ src/main/kotlin/com/atlassian/performance/tools/jiraperformancetests/api/RegressionResults.kt
@@ -2,7 +2,7 @@ package com.atlassian.performance.tools.jiraperformancetests.api
 
 import com.atlassian.performance.tools.report.api.result.CohortResult
 
-data class RegressionResults(
+class RegressionResults(
     val baseline: CohortResult,
     val experiment: CohortResult
 )